fix(test): repair phone normalization and argentina flow test rot, remove stale test artifact
