--- tests/inclusive-tax-test-results.txt
@@ -1,40 +1,0 @@
-Integration Test: Inclusive Tax Correctness
-[DB] Opening database at: /var/folders/0d/2kwkrk3j5j373ss9sbf8sd6m0000gn/T/flo-tax-incl-test-ddogJ0/flo.db
-[DB] Schema: v0 → v14
-[DB] Applying migration v1: initial_schema
-[DB] Install defaults loaded; first-run setup pending
-[DB] Migration v1 complete
-[DB] Applying migration v2: hash_plaintext_pins
-[DB] Migration v2 complete
-[DB] Applying migration v3: cloud_identity_and_outbox
-[DB] Migration v3 complete
-[DB] Applying migration v4: add_notes_limits_settings
-[DB] Migration v4 complete
-[DB] Applying migration v5: add_print_logs_table
-[DB] Migration v5 complete
-[DB] Applying migration v6: add_loyalty_settings
-[DB] Migration v6 complete
-[DB] Applying migration v7: add_discount_settings
-[DB] Migration v7 complete
-[DB] Applying migration v8: add_loyalty_index
-[DB] Migration v8 complete
-[DB] Applying migration v9: add_sequences_table
-[DB] Migration v9 complete
-[DB] Applying migration v10: fix_sequences_composite_key
-[DB] Migration v10 complete
-[DB] Applying migration v11: first_run_setup_uses_welcome_form
-[DB] Migration v11 complete
-[DB] Applying migration v12: fix_table_integer_ids
-[DB] Migration v12 complete
-[DB] Applying migration v13: fix_null_table_ids
-[DB] Migration v13 complete
-[DB] Applying migration v14: simplify_loyalty_settings
-[DB] Migration v14 complete
-[DB] integrity_check: ok
-[DB] foreign_key_check: clean
-[Auth] Generated new JWT secret for this install
-  ✓ order created
-  ✓ subtotal = ₹1000
-  ✓ total = ₹1000 (inclusive tax should not increase total)
-  ✓ discounted total = 900
-[DB] Database closed
